% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{calculate_loads}
\alias{calculate_loads}
\title{Emissions: calculate loads}
\usage{
calculate_loads(abimo_inpout, concentrations)
}
\arguments{
\item{abimo_inpout}{data.frame or SpatialPolygonsDataFrame with ABIMO input and
output as retrieved by \code{\link{postProcessABIMO}}}

\item{concentrations}{concentrations data frame as retrieved by
\code{\link{read_concentrations}}}
}
\value{
add calculated loads as additional colums  to abimo_inpout data.frame
or SpatialPolygonsDataFrame
}
\description{
The annual load is calculated with V x c. For for heavy metals
-> l/m2-year x ug/l = ug/m2-year;  for BOD/COD/TSS -> l/m2-year x mg/l = mg/m2-year
}
