% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{computeABIMOclimate}
\alias{computeABIMOclimate}
\title{abimo: compute climate}
\usage{
computeABIMOclimate(rawdir, fileName, skip, sep, dec, outAnnual, outSummer)
}
\arguments{
\item{rawdir}{rawdir}

\item{fileName}{fileName}

\item{skip}{skip}

\item{sep}{sep}

\item{dec}{dec}

\item{outAnnual}{outAnnual}

\item{outSummer}{outSummer}
}
\value{
???
}
\description{
read Climate Engine data and compute (source:
https://app.climateengine.org/climateEngine)
}
