% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_abimo_shares.R
\name{fix_abimo_shares}
\alias{fix_abimo_shares}
\title{Fix ABIMO shares}
\usage{
fix_abimo_shares(abimo)
}
\arguments{
\item{abimo}{abimo object}
}
\value{
fixed percental shares (PROBAU, PROVGU, STR_FLGES)
}
\description{
Fix ABIMO shares
}
