% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario_results.R
\name{get_abimo_stats}
\alias{get_abimo_stats}
\title{Get ABIMO Statistics}
\usage{
get_abimo_stats(abimo_inpout)
}
\arguments{
\item{abimo_inpout}{abimo_inpout}
}
\value{
tibble with columns "catchment_km2", "rainfall_cbm", "infiltration_cbm",
evapotrans_cbm and "vrr" (1 - runoff_cbm / rainfall_cbm)
}
\description{
Get ABIMO Statistics
}
