% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario_results.R
\name{get_scenario_results}
\alias{get_scenario_results}
\title{Get Scenario Results}
\usage{
get_scenario_results(paths)
}
\arguments{
\item{paths}{paths to directory containing all ABIMO scenario
results}
}
\value{
tibble
}
\description{
Get Scenario Results
}
