% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{makeFLGES}
\alias{makeFLGES}
\title{compute ABIMO variable FLGES (block area without street area)}
\usage{
makeFLGES(subcatchmSPobject)
}
\arguments{
\item{subcatchmSPobject}{subcatchmSPobject}
}
\value{
???
}
\description{
compute ABIMO variable FLGES (block area without street area)
}
