% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{makeOverlay}
\alias{makeOverlay}
\title{spatial overlay of subcatchments and raster holding information required by ABIMO}
\usage{
makeOverlay(
  rawdir,
  rasterData,
  subcatchmSPobject,
  overlayName,
  subcatchmNamesCol
)
}
\arguments{
\item{rawdir}{Path to data directory.}

\item{rasterData}{Name of raster file containing classified image.}

\item{subcatchmSPobject}{Spatial dataset containing subcatchment polygons
(ABIMO Blockteilflächen) (sp object type, R package sp).}

\item{overlayName}{Name of output overlay object.}

\item{subcatchmNamesCol}{Name of column in the attribute table of
subcatchmSPobject that contains the subcatchment identifiers. This
is used for naming the elements of the resulting list}
}
\value{
save overlay as .Rdata in directory "rawdir" with filename defined in
}
\description{
spatial overlay of subcatchments and raster holding information required by ABIMO
}
