% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{makePROBAU}
\alias{makePROBAU}
\title{compute ABIMO variable PROBAU}
\usage{
makePROBAU(rawdir, rasterData, overlayName, targetValue)
}
\arguments{
\item{rawdir}{rawdir}

\item{rasterData}{rasterData}

\item{overlayName}{overlayName}

\item{targetValue}{targetValue}
}
\value{
???
}
\description{
compute ABIMO variable PROBAU
}
