% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{makeVG}
\alias{makeVG}
\title{compute ABIMO variable VG (soil sealing percentage)}
\usage{
makeVG(rawdir, subcatchmSPobject, rasterData, targetValue)
}
\arguments{
\item{rawdir}{rawdir}

\item{subcatchmSPobject}{subcatchmSPobject}

\item{rasterData}{rasterData}

\item{targetValue}{targetValue}
}
\value{
???
}
\description{
based on online global land use data
}
