% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{padCODE}
\alias{padCODE}
\title{helper function: pad CODE column of ABIMO table}
\usage{
padCODE(string)
}
\arguments{
\item{string}{string with CODE identifier}
}
\value{
padded CODE identifier (with leading "0" depending of maximium character
length)
}
\description{
helper function: pad CODE column of ABIMO table
}
