% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{postProcessABIMO}
\alias{postProcessABIMO}
\title{abimo: postprocess}
\usage{
postProcessABIMO(path_input, path_output)
}
\arguments{
\item{path_input}{path of ABIMO input shapefile}

\item{abimo_output}{path of ABIMO output DBF file}
}
\value{

}
\description{
read dbf results file and joins with input shapefile
}
