% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgClass.R
\name{predictSurfClass}
\alias{predictSurfClass}
\title{apply model to predict surface type (roof, street, ...)}
\usage{
predictSurfClass(rawdir, modelName, image, predName)
}
\arguments{
\item{rawdir}{path to raw data directory}

\item{modelName}{modelName}

\item{image}{image}

\item{predName}{name of prediction file}
}
\value{
write raster in "rawdir" with file name "predName"
}
\description{
apply model to predict surface type (roof, street, ...)
}
