% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{read_concentrations}
\alias{read_concentrations}
\title{Emissions: read concentrations from OgRe database}
\usage{
read_concentrations(path)
}
\arguments{
\item{path}{path to OgRe database file "annual_mean_conc.csv"}
}
\value{
data frame with selected substances and column
}
\description{
imports data from OgRe database and selects relevant substances
for case study sites (Beijing, Jinxi) and calculates mean concentrations
over all structures (column: "mean"). In addition new columns (short_name,
unit_load, label_load) are created
}
