% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_results_beijing.R
\docType{data}
\name{scenario_results_beijing}
\alias{scenario_results_beijing}
\title{Results of ABIMO Scenario Analysis For Beijing}
\format{
A data.frame with 21 rows and 9 variables:
\describe{
\item{scenario_name}{name of scenario}
\item{catchment_km2}{sum of FLGES ans STR_FLGES (in square kilometers)}
\item{rainfall_cbm}{total rianfall in catchment  ABIMO (in cubicmeters/year)}
\item{runoff_cbm}{calculated runoff by ABIMO (in cubicmeter)}
\item{infiltration_cbm}{calculated infiltration by ABIMO (in cubicmeter/year)}
\item{evapotrans_cbm}{calculated evapotranspiration by ABIMO (in cubicmeter/year)}
\item{abimo_inpout}{tibble with ABIMO input/output (only water balance) }
\item{abimo_inpout_emissions}{tibble with ABIMO input/output (water balance + emissions) }
\item{vrr}{calculated volume rainfall retained (1-runoff_cbm/rainfall_cbm) }
}
}
\usage{
scenario_results_beijing
}
\description{
A dataset for ABIMO modelling results for Beijing case study
}
\keyword{datasets}
