% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{computeABIMOclimate}
\alias{computeABIMOclimate}
\title{abimo: compute climate}
\usage{
computeABIMOclimate(
  rawdir,
  file_inp,
  file_out,
  summer_month_start = 4,
  skip = 6,
  sep = "",
  dec = "."
)
}
\arguments{
\item{rawdir}{rawdir}

\item{file_inp}{name of input file}

\item{file_out}{name of output file to be written in "raw_dir"}

\item{summer_month_start}{number of month where summer half year starts
(default: 4)}

\item{skip}{skip (default: 6)}

\item{sep}{sep (default: '')}

\item{dec}{dec (default: '.')}
}
\value{
data frame with yearly summed measurements (summer half year,
total year sum) and also text file written to "raw_dir" with "out_file" name
}
\description{
read Climate Engine data and compute (source:
https://app.climateengine.org/climateEngine)
}
