% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{makeSTR_FLGES}
\alias{makeSTR_FLGES}
\title{compute ABIMO variable STR_FLGES}
\usage{
makeSTR_FLGES(
  rawdir,
  subcatchmSPobject,
  rasterData,
  overlayName,
  targetValue,
  mask
)
}
\arguments{
\item{rawdir}{rawdir}

\item{subcatchmSPobject}{subcatchmSPobject}

\item{rasterData}{rasterData}

\item{overlayName}{overlayName}

\item{targetValue}{targetValue}

\item{mask}{mask}
}
\value{
STR_FLGES
}
\description{
compute ABIMO variable STR_FLGES
}
