% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo.R
\name{postProcessABIMO}
\alias{postProcessABIMO}
\title{abimo: postprocess}
\usage{
postProcessABIMO(rawdir, nameABIMOin, nameABIMOout, ABIMOjoinedName)
}
\arguments{
\item{rawdir}{rawdir}

\item{nameABIMOin}{nameABIMOin}

\item{nameABIMOout}{nameABIMOout}

\item{ABIMOjoinedName}{ABIMOjoinedName}
}
\value{
write shapefile with name "ABIMOjoinedName" to "rawdir"
}
\description{
read dbf results file, join with input shapefile and write
output shapefile including ABIMO results
}
