% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgClass.R
\name{predictSurfClass}
\alias{predictSurfClass}
\title{apply model to predict surface type (roof, street, ...)}
\usage{
predictSurfClass(rawdir, modelName, image, predName)
}
\arguments{
\item{rawdir}{path to raw data directory}

\item{modelName}{Name of fitted random forest model saved by
\code{\link{buildClassMod}}}

\item{image}{Image to be classified}

\item{predName}{Name of output raster file (classified image)}
}
\value{
Writes raster in "rawdir" with file name "predName"
}
\description{
apply model to predict surface type (roof, street, ...)
}
